Starting libgdx implementation (using a dummy Egret implementation to get it compiling)

+++ dragonbones-core/src/main/java/com/dragonbones/model/TextureAtlasData.java
@@ -25,11 +25,11 @@ public abstract class TextureAtlasData extends BaseObject {
     /**
      * @private
      */
-    public float width;
+    public int width;
     /**
      * @private
      */
-    public float height;
+    public int height;
     /**
      * 贴图集缩放系数。
      *
